§4.3 C++ pass: drive SDNative + SDNativeTests warnings to zero on Release|x64
Two surfaces in one commit since the test project is the same toolchain and
the fixes are mechanical size_t→int narrowing.

  • Vendored: SDNative.vcxproj — per-file <DisableSpecificWarnings>4267;4334
        for 3rdparty/lodepng/lodepng.cpp. Project code stays under the gate.
  • SlabAllocator h+cpp — three call sites widen size_t into the alloc/addSlab
        uint32_t signature: explicit cast at the boundary.
  • ObjectCollection.cpp — totalMemory() narrows multiple .size()*sizeof(...)
        terms to uint32_t with explicit casts; snprintf("%d", ...size()) gets
        an (int) cast on the variadic argument (was C4477).
  • ShipDataSerializer.cpp — three int Len fields cast .size() at write.
  • SDNativeTests (TestFindNearby/TestQuadTree/TestSpatialGrid/SpatialSimUtils.h)
        — same shape; cast .size() at the int boundary of measureIterations,
        FoundCells::add, opt.MaxResults, total_operations.
  • NanoMesh submodule bumped to 5acc08b: Mesh_Obj.cpp lambda widens count
        to size_t; Mesh_Fbx.cpp LoadMaterial loop uses an int counter.

Verification: SDNative.vcxproj Release|x64 → 0 warnings (cpp-pass1-sdnative.log).
Solution Rebuild Release|x64 → 0 C/C++ warnings (cpp-pass1-full.log); only the
MSTest analyzer + NETSDK1206 noise remains, addressed in the next pass.

+++ SDNative/SDNativeTests/SpatialSimUtils.h
@@ -124,7 +124,7 @@ template<class Func> static void measureEachObj(const char* what, int iterations
         for (const MyGameObject& o : objects) { func(o); }
     }
     double e = t.elapsed_ms();
-    int total_operations = objects.size() * iterations;
+    int total_operations = (int)(objects.size() * iterations);
     printf("%s(%zu) x%d total: %.2fms  avg: %.3fus\n",
         what, objects.size(), iterations, e, (e / total_operations)*1000);
 }
